pyproject.toml
[project]
name = "foji-ai-api"
version = "0.1.0"
description = "Foji AI — Chat inference service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.6.0",
    "openai>=1.57.0",
    "google-genai>=1.0.0",
    "boto3>=1.35.0",
    "sse-starlette>=2.1.3",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
